+++ src/app/modules/evaluate/component/evaluate-settings/evaluate-settings.component.ts
@@ -15,7 +15,7 @@ export interface EvaluateUserSettings extends UserSettings {
   evaluateQueryOnline: boolean;
   evaluateQueryIndexed: ItemSearchIndexed;
   evaluateModifierRange: number;
-  evaluateModifierMaxRange: boolean;
+  evaluateModifierDisableMaxRange: boolean;
 }
 
 @Component({