+++ CHANGELOG.md
@@ -1,9 +1,10 @@
-## 0.5.1 (2020-01-xx)
+## 0.5.1 (2020-01-18)
 
 * add alternative text versions of stats to support negate or plural cases (#55, #68, #88, #96, #97)
 * add local stat support (#32)
+* add common pseudo filters (#7, #29)
 * add multiline stat support (#97)
-* add `Max Range Enabled` as settings (#83)
+* add `Disable Max Range` as settings (#83)
 * add regex caching for better parser performance
 * fix evaluate item frame changing size during loading (#83)
 * fix commands use wrong text (#95)